Refactor server architecture to replace apiRouterHandlerImpl
Replaced `apiRouterHandlerImpl` with `baseServer` to streamline server logic and improve extensibility. Updated related methods, middleware, and handler definitions across the codebase. Adjusted file structure and Makefile paths to reflect changes in server implementation.
