--- server/create_server.go
@@ -1,39 +1,0 @@
-package server
-
-import (
-	"github.com/gorilla/mux"
-	apicontext "github.com/softwareplace/http-utils/context"
-)
-
-func CreateApiRouter[T apicontext.Principal](topMiddlewares ...ApiMiddleware[T]) Api[T] {
-	router := mux.NewRouter()
-	router.Use(rootAppMiddleware[T])
-
-	api := &apiRouterHandlerImpl[T]{
-		router:                              router,
-		apiSecretKeyGeneratorResourceEnable: true,
-		loginResourceEnable:                 true,
-		contextPath:                         apiContextPath(),
-		port:                                apiPort(),
-	}
-
-	router.Use(api.errorHandlerWrapper)
-
-	for _, middleware := range topMiddlewares {
-		api.RegisterMiddleware(middleware, "")
-	}
-	return api.NotFoundHandler()
-}
-
-func CreateApiRouterWith[T apicontext.Principal](router mux.Router) Api[T] {
-	router.Use(rootAppMiddleware[T])
-	api := &apiRouterHandlerImpl[T]{
-		router:                              &router,
-		apiSecretKeyGeneratorResourceEnable: true,
-		loginResourceEnable:                 true,
-		contextPath:                         apiContextPath(),
-		port:                                apiPort(),
-	}
-
-	return api.NotFoundHandler()
-}
