--- server/router/router.go
@@ -1,1 +1,0 @@
-package router
