+++ server/routers_impl.go
@@ -7,14 +7,14 @@ import (
 	"strings"
 )
 
-func (a *apiRouterHandlerImpl[T]) PublicRouter(handler ApiContextHandler[T], path string, method string) Api[T] {
+func (a *baseServer[T]) PublicRouter(handler ApiContextHandler[T], path string, method string) Api[T] {
 	a.Add(handler, path, method)
 	combinedKey := method + "::" + a.contextPath + path
 	principal.AddOpenPath(combinedKey)
 	return a
 }
 
-func (a *apiRouterHandlerImpl[T]) Add(handler ApiContextHandler[T], path string, method string, requiredRoles ...string) Api[T] {
+func (a *baseServer[T]) Add(handler ApiContextHandler[T], path string, method string, requiredRoles ...string) Api[T] {
 
 	handlerPath := strings.TrimSuffix(a.contextPath, "/") + "/" + strings.TrimPrefix(path, "/")
 
@@ -27,37 +27,37 @@ func (a *apiRouterHandlerImpl[T]) Add(handler ApiContextHandler[T], path string,
 	return a
 }
 
-func (a *apiRouterHandlerImpl[T]) Get(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
+func (a *baseServer[T]) Get(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
 	a.Add(handler, path, "GET", requiredRoles...)
 	return a
 }
 
-func (a *apiRouterHandlerImpl[T]) Post(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
+func (a *baseServer[T]) Post(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
 	a.Add(handler, path, "POST", requiredRoles...)
 	return a
 }
 
-func (a *apiRouterHandlerImpl[T]) Put(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
+func (a *baseServer[T]) Put(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
 	a.Add(handler, path, "PUT", requiredRoles...)
 	return a
 }
 
-func (a *apiRouterHandlerImpl[T]) Delete(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
+func (a *baseServer[T]) Delete(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
 	a.Add(handler, path, "DELETE", requiredRoles...)
 	return a
 }
 
-func (a *apiRouterHandlerImpl[T]) Patch(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
+func (a *baseServer[T]) Patch(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
 	a.Add(handler, path, "PATCH", requiredRoles...)
 	return a
 }
 
-func (a *apiRouterHandlerImpl[T]) Options(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
+func (a *baseServer[T]) Options(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
 	a.Add(handler, path, "OPTIONS", requiredRoles...)
 	return a
 }
 
-func (a *apiRouterHandlerImpl[T]) Head(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
+func (a *baseServer[T]) Head(handler ApiContextHandler[T], path string, requiredRoles ...string) Api[T] {
 	a.Add(handler, path, "HEAD", requiredRoles...)
 	return a
 }
