+++ server/server_impl.go
@@ -1,0 +1,71 @@
+package server
+
+import (
+	"github.com/gorilla/mux"
+	errorhandler "github.com/softwareplace/http-utils/error"
+	"github.com/softwareplace/http-utils/security/principal"
+	"net/http"
+	"strings"
+)
+
+func (a *baseServer[T]) RegisterCustomMiddleware(middleware func(next http.Handler) http.Handler) Api[T] {
+	a.router.Use(middleware)
+	return a
+}
+
+func (a *baseServer[T]) ErrorHandler(handler errorhandler.ApiHandler[T]) Api[T] {
+	a.errorHandler = handler
+	return a
+}
+
+func (a *baseServer[T]) LoginResource(loginService LoginService[T]) Api[T] {
+	a.loginService = loginService
+	if a.loginResourceEnable {
+		a.PublicRouter(a.Login, "login", "POST")
+	}
+	return a
+}
+
+func (a *baseServer[T]) ApiKeyGeneratorResource(apiKeyGeneratorService ApiKeyGeneratorService[T]) Api[T] {
+	a.apiKeyGeneratorService = apiKeyGeneratorService
+	if a.apiSecretKeyGeneratorResourceEnable {
+		a.Post(a.ApiKeyGenerator, "api-key/generate", "POST", strings.Join(apiKeyGeneratorService.RequiredScopes(), " "))
+	}
+	return a
+}
+
+func (a *baseServer[T]) SecretKeyGeneratorResourceEnabled(enable bool) Api[T] {
+	a.apiSecretKeyGeneratorResourceEnable = enable
+	return a
+}
+
+func (a *baseServer[T]) LoginResourceEnabled(enable bool) Api[T] {
+	a.loginResourceEnable = enable
+	return a
+}
+
+func (a *baseServer[T]) PrincipalService(service principal.Service[T]) Api[T] {
+	a.principalService = service
+	if a.principalService != nil {
+		a.router.Use(a.hasResourceAccess)
+	}
+	return a
+}
+
+func (a *baseServer[T]) ServeHTTP(w http.ResponseWriter, req *http.Request) {
+	a.router.ServeHTTP(w, req)
+}
+
+func (a *baseServer[T]) Router() *mux.Router {
+	return a.router
+}
+
+func (a *baseServer[T]) RouterHandler(handler RouterHandler) Api[T] {
+	handler(a.router)
+	return a
+}
+
+func (a *baseServer[T]) EmbeddedServer(handler func(Api[T])) Api[T] {
+	handler(a)
+	return a
+}
